Modify seed-from-info feature: More than cbase and rbase
This is more like what I had in mind: all attributes that are
present, for a given setup, in the seed info file basically always get
inherited in the resulting merged test.info file.
Attribute values from the seed info file have priority, when
an attribute is present both in the seed info file (for a given setup)
and the implied attribute set (which is based on .suite and .yaml files and
various flags) that would be used in the absence of a seed info file,
the value from the seed info file wins out.

Note that the set of setups for which the resulting "merged" info file
has nodes is still that defined by the .suite files. In other words,
the set of setups from the seed info file does not really get merged
with that from the seed info; merging (set union) only takes place at
the level of attribute sets for any setup that is in the intersection
of the set of setups defined by .suite files and the set of setups
existing in the seed info file.